Hide the lang menu when not needed
Define the block trmenu in menu structure of the base template. In both articles and pages, define this block. Inside the block, we first check if there are any translations at all. If there are none, nothing happens. If there are some, the menu item for language selection is inserted. Also, the translations macro gets called